mock classes to test.py; small test updates
